v0/requirements.txt
openai
omegaconf
hydra-core
pydantic
numpy
matplotlib
Pillow
pyyaml
